support labels for blank values
This adds support for blank labels on nullable choices fields by using
the `__empty__` value to set the label. This feature is deeply hidden in
the Django documentation in the bottom of the section around
["Enumeration Types"](https://docs.djangoproject.com/en/5.2/ref/models/fields/#enumeration-types):

> Because an enumeration with a concrete data type requires all values
> to match the type, overriding the blank label cannot be achieved by
> creating a member with a value of None. Instead, set the __empty__
> attribute on the class:
>
> ```
>     class Answer(models.IntegerChoices):
>         NO = 0, _("No")
>         YES = 1, _("Yes")
>
>         __empty__ = _("(Unknown)")
> ```

The empty value is included in `<enum_type>.choices` as `(None,
"label")` but can be avoided with `[(x.value, x.label) for x in <enum_type>]`.